libDispatch building completely on windows when using mingw
git-svn-id: http://opensource.mlba-team.de/svn/xdispatch/trunk@75 9e4c620c-c391-4b63-a7fb-4924991ea895

+++ libkqueue/src/common/debug.h
@@ -59,7 +59,7 @@ extern char *DEBUG_IDENT;
 #  define dbg_lasterror(str)     do {                                \
     if (DEBUG_LEVEL)                                                      \
       fprintf(stderr, "%s: [%d] %s(): %s: (LastError=%d)\n",        \
-              DEBUG_IDENT, THREAD_ID, __func__, str, GetLastError());            \
+              DEBUG_IDENT, THREAD_ID, __func__, str, (int)GetLastError());            \
 } while (0)
 # else
 #  define dbg_lasterror(str)     ;
